% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulize.R
\name{formula_helper}
\alias{formula_helper}
\title{Convert a formula and data frame to a design matrix and response vector}
\usage{
formula_helper(formula, data, ...)
}
\arguments{
\item{formula}{Formula specifying design matrix.}

\item{data}{Data frame in which to evaluate formula.}

\item{...}{Included so we can check if an \code{intercept} was specified.}
}
\value{
A list with three named entries: a data matrix \code{X},
  a response vector (potentially a matrix later on) \code{y}, and
  a logical \code{intercept} indicating if the formula specified that
  an intercept should be added to the design matrix.
}
\description{
Note that this function is not exported.
}
