% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulize.R
\name{formulize}
\alias{formulize}
\title{Create a formula interface to a modelling method}
\usage{
formulize(f)
}
\arguments{
\item{f}{A function with a matrix/vector interface.}
}
\value{
The same function, wrapped to have a formula/dataframe interface.
}
\description{
Create a formula interface to a modelling method
}
\examples{

library(glmnet)

glmnet_form <- formulize(cv.glmnet)

model <- glmnet_form(mpg ~ drat * hp - 1, mtcars)
predict(model, head(mtcars))

}
