% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulize.R
\name{genericize}
\alias{genericize}
\title{A good day to title hard}
\usage{
genericize(f, fname = NULL)
}
\arguments{
\item{f}{more stuff}

\item{fname}{so much good documentation}
}
\description{
TODO: calls like glmnet::cv.glmnet break function name discovery, maybe other things
TODO: don't call genericize twice, call it once for side-effects
TODO: don't call genericize on things that already have a nice formula interface
}
\details{
Title
}
