% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulize.R
\name{predict.wrapped}
\alias{predict.wrapped}
\title{Predict on new data form a \code{wrapped} object}
\usage{
\method{predict}{wrapped}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{Object of class \code{wrapped} with a subclass with a
valid \code{predict} method.}

\item{newdata}{Data to predict on.}

\item{...}{Additional arguments to pass to the subclass predict method.}
}
\value{
Predictions from subclass predict method.
}
\description{
Predict on new data form a \code{wrapped} object
}
\seealso{
Other chirunconf: \code{\link{formulize}},
  \code{\link{genericize}}
}
\concept{chirunconf}
