% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulize.R
\name{genericize}
\alias{genericize}
\title{Turn a modelling function into an S3 modelling generic}
\usage{
genericize(f, fname = NULL)
}
\arguments{
\item{f}{Modelling function to turn into a generic.}

\item{fname}{Optional name of new generic. Useful because otherwise the
created generic masks the original modelling function.}
}
\value{
Nothing.
}
\description{
Call this function for its side effects: (1) creating an S3 generic of the
passed modelling function, (2) creating a default method that's exactly the
X/y interface to formulize, (3) creating a formula method with formulize.
}
\seealso{
Other chirunconf: \code{\link{formulize}},
  \code{\link{predict.wrapped}}
}
\concept{chirunconf}
