% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulize.R
\name{predict.wrapped}
\alias{predict.wrapped}
\title{TODOCUMENTDO}
\usage{
\method{predict}{wrapped}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{doc doc}

\item{newdata}{more doc}

\item{...}{woops}
}
\description{
TODOCUMENTDO
}
